refactor(notifications): extract getNovuClient into dedicated module
Moves getNovuClient into novu-client.ts to avoid pulling the heavy
notifications barrel (Authenticator, DataSourceResource, etc.) into
workflow files, which would otherwise create an import cycle.
